Add user_id on categories and schedules

+++ app/api/labels/route.ts
@@ -1,10 +1,18 @@
 import { NextResponse, NextRequest } from "next/server";
 
 import connectionPool from "@/lib/db";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/lib/auth";
+import { getValidatedSession } from "@/lib/session";
 
 export async function GET(request: Request) {
     try {
-        const query = 'SELECT * FROM mysched.categories';
+        
+        const userId = await getValidatedSession();
+
+
+        const query = `SELECT * FROM mysched.categories WHERE user_id IS NULL OR user_id = ${userId}`;
+        // const query = 'SELECT * FROM mysched.categories';
 
         const [results] = await connectionPool.execute(query);
 
@@ -34,11 +42,13 @@ export async function POST(request: Request) {
                 { status: 400 }
             );  
         }
-    
-        // insert data into the database
-        const query = `INSERT INTO mysched.categories (name, color) VALUES (?, ?)`;
 
-        const [result] = await connectionPool.execute(query, [name, color]);
+        const userId = await getValidatedSession();
+
+        // insert data into the database
+        const query = `INSERT INTO mysched.categories (name, color, user_id) VALUES (?, ?, ?)`;
+
+        const [result] = await connectionPool.execute(query, [name, color, userId]);
 
         // Respond with success
         return NextResponse.json({ success: true });
@@ -62,9 +72,11 @@ export async function DELETE(request: NextRequest) {
             );
         }
 
+        const userId = await getValidatedSession();
+
         // execute delete query
-        const query = `DELETE FROM mysched.categories WHERE id = ?`;
-        const [result] = await connectionPool.execute(query, [id]);
+        const query = `DELETE FROM mysched.categories WHERE id = ? AND user_id = ?`;
+        const [result] = await connectionPool.execute(query, [id, userId]);
 
         if (!result) {
             return NextResponse.json(
