+++ app/api/schedule/route.ts
@@ -1,9 +1,16 @@
 import { NextResponse, NextRequest } from "next/server";
 import { formatDate, formatToZonedTime } from "@/lib/utils";
 import connectionPool from "@/lib/db";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/lib/auth";
+import { getValidatedSession } from "@/lib/session";
 
 export async function POST(request: Request) {
     try {
+        const session = await getServerSession(authOptions);
+        if(!session || !session.user?.id) {
+            return new Response(JSON.stringify({ error: "Unauthorized" }));
+        }
         // parse the request body
         const body = await request.json();
 
@@ -24,8 +31,8 @@ export async function POST(request: Request) {
         const formattedEndTime = formatToZonedTime(end_time);
     
         // insert the data into the database
-        const query = `INSERT INTO mysched.schedules (title, description, date, start_time, end_time, color, category_id) VALUES (?, ?, ?, ?, ?, ?, ?)`;
-        const [result] = await connectionPool.execute(query, [title, description, formattedDate, formattedStartTime, formattedEndTime, color, category_id]);
+        const query = `INSERT INTO mysched.schedules (title, description, date, start_time, end_time, color, category_id, user_id) VALUES (?, ?, ?, ?, ?, ?, ?, ?)`;
+        const [result] = await connectionPool.execute(query, [title, description, formattedDate, formattedStartTime, formattedEndTime, color, category_id, session.user.id]);
 
         // respond with success
         return NextResponse.json({ success: true });
@@ -43,6 +50,8 @@ export async function DELETE(request: NextRequest) {
         const { searchParams } = new URL(request.url);
         const id = searchParams.get("id");
 
+        const userId = await getValidatedSession();
+
         if (!id) {
             return NextResponse.json(
                 {error: "Schedule ID is required."},
@@ -51,8 +60,8 @@ export async function DELETE(request: NextRequest) {
         }
 
         // execute delete query
-        const query = `DELETE FROM mysched.schedules WHERE id = ?`;
-        const [result] = await connectionPool.execute(query, [id]);
+        const query = `DELETE FROM mysched.schedules WHERE id = ? AND schedules.user_id = ?`;
+        const [result] = await connectionPool.execute(query, [id, userId]);
 
         if (!result) {
             return NextResponse.json(
